+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - removed Gemini CLI provider due to end of support for subscription from Google
 - correctly placing skill messages after the user prompt when used from custom commands
 - added skillsToActivate to onAgentStarted extension event to allow activating skills
+- added noProxy support to network settings
 
 ## [0.70.0]